pyproject.toml
[project]
name = "tomarkdown"
version = "0.1.0"
description = "Convert documents (DOCX, XLSX, PDF) to Markdown"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markitdown[all]>=0.0.1",
    "pymupdf4llm>=0.0.1",
]

[project.scripts]
tomarkdown = "tomarkdown.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "python-docx>=1.0.0",
    "reportlab>=4.0.0",
]

[tool.uv]
package = true
